chore: split Input component
